serial: 8250: powerpc: Make gem5 uart port global
As seen in commit 811c823877c4, the only source of an external
interrupt in gem5 is the 8250 UART. Since the simulator does
not currently model an interrupt controller, there is no way
to programatically determine the source. This makes the only
uart_port instance globally visible so that it is accessible
to the arch-specific interrupt handling code.

Fixes: 811c823877c4 ("powerpc: powernv: Add hacked up support for 8250 serial console")
Signed-off-by: Sandipan Das <sandipan@linux.ibm.com>